fixes#5

+++ src/Main.java
@@ -5,7 +5,7 @@ public class Main {
         int distanceKm = 100;
         float fuelConsumptionAtADist = 145.8f;
         float fuelAvailable = 125.9f;
-        float allowableDistanceKm = distanceKm * fuelConsumptionAtADist / fuelAvailable;
+        int allowableDistanceKm = (int) (distanceKm * fuelConsumptionAtADist / fuelAvailable);
         System.out.println(allowableDistanceKm+"км");
     }
 }
